# docs/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: Enclii Platform API
  description: |
    Open source DevOps platform API for deploying and managing containerized applications.

    ## Overview

    Enclii Switchyard is the control plane API that manages projects, services, deployments,
    domains, teams, and infrastructure. It provides a comprehensive REST API with 107+ endpoints
    covering the complete application lifecycle.

    ## Authentication

    Most endpoints require JWT authentication. There are two authentication modes:

    - **OIDC Mode** (Production): Uses Janua SSO with RS256 JWT validation
    - **Local Mode** (Development): Uses self-signed JWT with password-based auth

    Include the token in the `Authorization` header as `Bearer <token>`.

    ## Rate Limiting

    - Auth endpoints: 10 requests/minute (5 for register/login)
    - Standard endpoints: 100 requests/minute
    - Build endpoints: 20 requests/minute

    ## Environments

    Services can be deployed to multiple environments:
    - `dev` - Development (default)
    - `staging` - Pre-production testing
    - `prod` - Production

    ## WebSocket Endpoints

    Real-time log streaming is available via WebSocket:
    - `ws://api.enclii.dev/v1/services/{id}/logs/stream` - Service logs
    - `ws://api.enclii.dev/v1/deployments/{id}/logs/stream` - Deployment logs
    - `ws://api.enclii.dev/v1/services/{id}/builds/{build_id}/logs/stream` - Build logs

  version: 1.0.0
  contact:
    name: MADFAM Engineering
...
servers:
...
paths:
...
  /projects:
    get:
      summary: List projects
      description: Get all projects accessible to the authenticated user.
      tags: [projects]
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      summary: Create project
      description: Create a new project. Requires admin role.
      tags: [projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '409':
          description: Project slug already exists
  /projects/{slug}:
    get:
      summary: Get project
      description: Get project details by slug.
      tags: [projects]
      operationId: getProject
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: my-project
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found

  # ============================================
  # CI RUNNER CONFIGURATION
  # ============================================
...
  /v1/projects/{slug}/junctions:
    post:
      summary: Create junction
      description: Creates a routing rule with Cloudflare tunnel + DNS provisioning
      tags: [junctions]
      operationId: createJunction
      parameters:
        - $ref: '#/components/parameters/projectSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id, domain]
              properties:
                service_id:
                  type: string
                  format: uuid
                domain:
                  type: string
                  description: "e.g., api.example.com"
                path:
                  type: string
                  default: "/"
                protocol:
                  type: string
                  enum: [http, https, grpc]
                  default: https
                tls:
                  type: object
                  properties:
                    enabled:
                      type: boolean
                      default: true
                    issuer:
                      type: string
                      enum: [letsencrypt-prod, letsencrypt-staging, custom, selfsigned]
                      default: letsencrypt-prod
                    cert_secret:
                      type: string
                    min_version:
                      type: string
                      default: "1.2"
                    force_redirect:
                      type: boolean
                      default: true
      responses:
        '201':
          description: Junction created and infrastructure provisioned
        '400':
          description: Invalid request
        '404':
          description: Project not found
        '409':
          description: Junction for this domain+path already exists
    get:
      summary: List junctions
      description: Lists all routing rules for a project
      tags: [junctions]
      operationId: listJunctions
      parameters:
        - $ref: '#/components/parameters/projectSlug'
      responses:
        '200':
          description: List of junctions
  /v1/junctions/{id}:
    get:
      summary: Get junction
      tags: [junctions]
      operationId: getJunction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Junction details
        '404':
          description: Junction not found
    delete:
      summary: Delete junction
      description: Deletes a junction and removes tunnel route + DNS record
      tags: [junctions]
      operationId: deleteJunction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Junction deleted and infrastructure cleaned up
        '404':
          description: Junction not found

  # ============================================
  # P2.6 - v-number deployment resolution
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or OIDC callback
  parameters:
    limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 50
        maximum: 100
    offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        default: 0
    projectSlug:
      name: slug
      in: path
      required: true
      description: Project slug
      schema:
        type: string
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 2
          maxLength: 50

    # ===== Environments =====
...
security:
  - bearerAuth: []